pigop-backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ── Cache & Queue ─────────────────────────────────────────────────────────────
redis==5.0.1
celery==5.3.4

# ── Google Cloud ──────────────────────────────────────────────────────────────
google-cloud-storage==2.10.0
google-genai>=1.0.0          # Fase 3: Gemini AI SDK (nuevo — reemplaza google-generativeai)

# ── PDF / OCR (Fase 3) ────────────────────────────────────────────────────────
pypdf>=4.0.0                 # Extracción texto PDF nativo

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.25.2

# ── Utils ─────────────────────────────────────────────────────────────────────
python-dateutil==2.8.2

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── PDF Generation ────────────────────────────────────────────────────────────
reportlab==4.0.7

# ── DOCX Generation ──────────────────────────────────────────────────────────
python-docx==1.1.0

# ── QR Code & Firma ─────────────────────────────────────────────────────────
qrcode[pil]>=7.4

# ── Criptografía e.firma (AES-256-GCM + X.509 + RSA) ──────────────────────
cryptography>=42.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0
